+++ examples/mrminio/main.go
@@ -5,13 +5,12 @@ import (
 
 	"github.com/mondegor/go-webcore/mrlib"
 	"github.com/mondegor/go-webcore/mrlog"
-	"github.com/mondegor/go-webcore/mrlog/mrlogbase"
 
 	"github.com/mondegor/go-storage/mrminio"
 )
 
 func main() {
-	logger := mrlogbase.New(mrlog.TraceLevel)
+	logger := mrlog.New(mrlog.TraceLevel)
 	ctx := mrlog.WithContext(context.Background(), logger)
 
 	logger.Info().Msg("Create minio S3 connection")