+++ examples/mrsql/builderwhere/main.go
@@ -3,14 +3,13 @@ package main
 import (
 	"github.com/google/uuid"
 	"github.com/mondegor/go-webcore/mrlog"
-	"github.com/mondegor/go-webcore/mrlog/mrlogbase"
 	"github.com/mondegor/go-webcore/mrtype"
 
 	"github.com/mondegor/go-storage/mrpostgres"
 )
 
 func main() {
-	logger := mrlogbase.New(mrlog.TraceLevel)
+	logger := mrlog.New(mrlog.TraceLevel)
 	bw := mrpostgres.NewSQLBuilderWhere()
 
 	where := bw.JoinOr(