+++ mrpostgres/query_row.go
@@ -1,23 +1,29 @@
 package mrpostgres
 
 import (
+	"errors"
+
 	"github.com/jackc/pgx/v5"
+	"github.com/mondegor/go-webcore/mrcore"
 )
 
 type (
 	queryRow struct {
 		row pgx.Row
-		err error
 	}
 )
 
 // Scan - comment method.
 func (qr *queryRow) Scan(dest ...any) error {
-	if qr.err != nil {
-		return qr.err
-	}
-
 	if err := qr.row.Scan(dest...); err != nil {
+		if errors.Is(err, pgx.ErrNoRows) {
+			return mrcore.ErrStorageNoRowFound.Wrap(err)
+		}
+
+		if errors.Is(err, pgx.ErrTooManyRows) {
+			return mrcore.ErrStorageFetchDataFailed.Wrap(err)
+		}
+
 		return wrapError(err)
 	}